// apps/web/src/api/orders.gen.ts
export type OrderOverviewDtoStatus = typeof OrderOverviewDtoStatus[keyof typeof OrderOverviewDtoStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OrderOverviewDtoStatus = {
  CREATED: 'CREATED',
  STOCK_RESERVED: 'STOCK_RESERVED',
  PAYMENT_PENDING: 'PAYMENT_PENDING',
  PAID: 'PAID',
  READY_FOR_PICKING: 'READY_FOR_PICKING',
  COMPLETED: 'COMPLETED',
  STOCK_RESERVATION_FAILED: 'STOCK_RESERVATION_FAILED',
  PAYMENT_FAILED: 'PAYMENT_FAILED',
  STOCK_RELEASED: 'STOCK_RELEASED',
  CANCELLED: 'CANCELLED',
} as const;

export interface OrderOverviewDto {
  orderId?: string;
  customerId?: string;
  status?: OrderOverviewDtoStatus;
  totalAmount?: number;
  itemCount?: number;
  createdAt?: string;
  updatedAt?: string;
}